Correccion de vista Buscar

+++ BibliotecaMovil.Shared/DTOs/ResenaDto.cs
@@ -7,7 +7,7 @@ public class ResenaDto
     public int IdLibro { get; set; }
     public string? Titulo { get; set; }
     public string Contenido { get; set; } = string.Empty;
-    public int? Puntuacion { get; set; }
+    public decimal? Puntuacion { get; set; }
     public DateTime FechaCreacion { get; set; } = DateTime.Now;
     public DateTime? FechaModificacion { get; set; }
 }
